Split code into separate packages
