+++ main.go
@@ -1,7 +1,6 @@
 package main
 
 import (
-	"encoding/json"
 	"errors"
 	"fmt"
 	"html/template"
@@ -9,103 +8,25 @@ import (
 	"net/http"
 	"os"
 	"strconv"
-	"time"
 
+	"github.com/saaste/memento-mori/calendar"
+	"github.com/saaste/memento-mori/config"
 	"github.com/saaste/memento-mori/utils"
 )
-
-type AppConfig struct {
-	Birthday       utils.Date
-	LifeExpectancy int
-	Events         []Event
-}
-
-type Event struct {
-	Date  utils.Date
-	Title string
-	Label string
-}
-
-type Week struct {
-	Class string
-	Event Event
-}
-
-type Year struct {
-	Year  int
-	Weeks []Week
-}
 
 type IndexData struct {
 	Birthday       utils.Date
 	LifeExpectancy int
-	Years          []Year
-}
-
-func getWeeks(year int, month int, config AppConfig) []Week {
-
-	now := time.Now()
-	daysInMonth := utils.DaysInMonth(time.Date(year, time.Month(month), 1, 0, 0, 0, 0, time.UTC))
-	daysInSquare := daysInMonth / 4
-
-	weeks := make([]Week, 0)
-	for w := 0; w < 4; w++ {
-		startOfSquare := time.Date(year, time.Month(month), daysInSquare*w+1, 0, 0, 0, 0, time.UTC)
-		endOfSquare := time.Date(year, time.Month(month), daysInSquare*(w+1), 0, 0, 0, 0, time.UTC)
-		if w == 3 && (endOfSquare.Day() != daysInMonth || endOfSquare.Month() != startOfSquare.Month()) {
-			endOfSquare = time.Date(year, time.Month(month), daysInMonth, 0, 0, 0, 0, time.UTC)
-		}
-
-		class := ""
-		if startOfSquare.Before(config.Birthday.Time) {
-			class = "hidden"
-		} else if endOfSquare.Before(now) {
-			class = "past"
-		}
-
-		newWeek := Week{
-			Class: class,
-			Event: Event{},
-		}
-
-		for _, event := range config.Events {
-			if event.Date.Equal(startOfSquare) || event.Date.Equal(endOfSquare) || (event.Date.After(startOfSquare) && event.Date.Before(endOfSquare)) {
-				year = event.Date.Year()
-				newWeek.Event = event
-				break
-			}
-		}
-
-		weeks = append(weeks, newWeek)
-	}
-	return weeks
-}
-
-func getYears(yearOfBirth int, lifeExpectancy int, config AppConfig) []Year {
-	firstYear := yearOfBirth
-	lastYear := yearOfBirth + lifeExpectancy
-
-	years := make([]Year, 0)
-	for currentYear := firstYear; currentYear <= lastYear; currentYear++ {
-		weeks := make([]Week, 0)
-		for currentMonth := 1; currentMonth < 12; currentMonth++ {
-			weeks = append(weeks, getWeeks(currentYear, currentMonth, config)...)
-		}
-		years = append(years, Year{
-			Year:  currentYear,
-			Weeks: weeks,
-		})
-	}
-	return years
+	Years          []calendar.Year
 }
 
 func handleRoot(w http.ResponseWriter, r *http.Request) {
-	config, err := readConfig()
+	config, err := config.ReadConfig()
 	if err != nil {
 		log.Fatalf("ERROR: reading config.json failed: %s\n", err)
 	}
 
-	years := getYears(config.Birthday.Year(), config.LifeExpectancy, config)
+	years := calendar.GetYears(config.Birthday.Year(), config.LifeExpectancy, config)
 	tmpl := template.Must(template.ParseFiles("./templates/index.html"))
 	data := IndexData{
 		Birthday:       config.Birthday,
@@ -113,21 +34,6 @@ func handleRoot(w http.ResponseWriter, r *http.Request) {
 		Years:          years,
 	}
 	tmpl.Execute(w, data)
-}
-
-func readConfig() (AppConfig, error) {
-	var config AppConfig
-
-	f, err := os.ReadFile("./config.json")
-	if err != nil {
-		return config, err
-	}
-	err = json.Unmarshal([]byte(f), &config)
-	if err != nil {
-		return config, err
-	}
-	return config, nil
-
 }
 
 func main() {
